added lrs tso1 to lrsslitless tests
